adds profile controller
