+++ controllers/index.js
@@ -1,5 +1,6 @@
 module.exports = {
     auth: require('./auth'),
     message: require('./message'),
-    nudge: require('./nudge')
+    nudge: require('./nudge'),
+    profile: require('./profile')
 };
