+++ views/profile.ejs
@@ -34,7 +34,7 @@
       <ul>
         <% user.nudges.forEach(nudge => { %>
         <li>
-          <%= nudge.taskName %> 
+          <%= nudge.taskName %>
         </li>
         <% }) %>
